implement main receive logic of the distance vector routing
For each destination in the vector we check if the destination is know
to us or not. If the destination is unknown we add the destination to
the routing table. If the destination is know we check if the we are the
destination. If we are the destination we drop don't do anything. We
only accept updates about other destinations because we know already
everything we need to know about us as destination. If the metric to the
destination improves we update the destination. If we get an update for
a destination from the same link we received information about it before
we update the destination as well. This helps us learn bad news.