don't check for destinations in vector that are local
We don't need to check for destinations that are local to the node in
the vector because we never will get information about local
destinations via the loop back interface. We will only get updates from our
neighbours about us via the external interfaces. we do not update our
routing table in this case because the cost is not better and the link
of the know destination is not equal to the interface we receive the
information from.